Revert "Merge branch 'Function-TryCatch' of github.com:SchulichUAV/AutopilotDevelopment into Function-TryCatch"
This reverts commit 2648a9b0317a4adf0b779dfaba0a963400b37659, reversing
changes made to 3a6b639857e93ae4779e78b24913636e7d295155.
