+++ Plane/plane.py
@@ -48,11 +48,6 @@ class Plane:
     def get_gps_raw(self):
         return plane_system_state.receive_gps_raw(self.vehicle_connection)
     
-    def get_lat_long(self):
-        lat_long_data = plane_system_state.receive_lat_long(self.vehicle_connection)
-        if lat_long_data:
-            print(f"Latitude: {lat_long_data['latitude']}, Longitude: {lat_long_data['longitude']}")
-        return lat_long_data
     ## from system state, we still need gps status, scaled imu, utm_global_position and wind_cov
 
     ## takeoff configurations
@@ -74,3 +69,14 @@ class Plane:
 
     def set_rtl_radius(self, rtl_loiter_radius):
         plane_waypoint.set_waypoint_rtl_loiter_radius(self.vehicle_connection, rtl_loiter_radius)
+
+    
+if __name__ == '__main__':
+    ## establish connection to the plane
+    plane = Plane(vehicle_connection='udpin:127.0.0.1:14550')
+
+    ## arm the vehicle
+    plane.arm_vehicle()
+
+    ## example call
+    plane.set_speed(15)
